requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==44.0.3
docutils==0.21.2
filelock==3.18.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.7
jnius==1.1.0
keyring==25.6.0
numpy==2.2.5
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
protobuf==6.30.2
pyOpenSSL==25.0.0
PyYAML==6.0.2
PyYAML==6.0.2
redis==6.1.0
Sphinx==8.2.3
streamlit==1.44.0
thread==2.0.5
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
ultralytics==8.3.98
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
